workspaces/telenordeste_integration/requirements.txt
# Notion Integration
notion-client==2.2.1

# Google Calendar Integration
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3

# Date/Time handling
python-dateutil==2.8.2
